Enable user to update profile info with RF

+++ job/views.py
@@ -1,16 +1,15 @@
-from django.shortcuts import render
+from django.shortcuts import render, get_object_or_404
 from django.db.models import Avg, Max, Min, Count
 
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.pagination import PageNumberPagination
+from rest_framework.permissions import IsAuthenticated
 
 from .serializers import JobSerializer
 from .models import Job
 from .filters import JobsFilter
-
-from django.shortcuts import get_object_or_404, get_list_or_404
 
 
 # Get job list
@@ -47,7 +46,10 @@ def getJob(request, pk):
 
 # Post new job
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def newJob(request):
+
+    request.data['user'] = request.user
     data = request.data
 
     job = Job.objects.create(**data)
@@ -58,8 +60,16 @@ def newJob(request):
 
 # Update job by id
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def updateJob(request, pk):
+
     job = get_object_or_404(Job, id=pk)
+
+    if job.user != request.user:
+        return Response(
+            {'message': 'You are not allowed to update this job'},
+            status=status.HTTP_403_FORBIDDEN
+        )
 
     job.title = request.data['title']
     job.description = request.data['description']
@@ -81,8 +91,15 @@ def updateJob(request, pk):
 
 # Delete job by id
 @api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def deleteJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+
+    if job.user != request.user:
+        return Response(
+            {'message': 'You are not allowed to update this job'},
+            status=status.HTTP_403_FORBIDDEN
+        )
 
     job.delete()
     return Response({'message: Item deleted'}, status=status.HTTP_200_OK)
